+++ src/services/api.js
@@ -1,7 +1,7 @@
 import axios from 'axios';
 
-const baseURL = 'https://fakestoreapi.com'; //서버 연결전까지 사용할 fakeApi
-// const baseURL = 'https://trendy-healthy-backend.store'; //서버 연결 후 사용할 APi
+// const baseURL = 'https://fakestoreapi.com'; //서버 연결전까지 사용할 fakeApi
+const baseURL = 'https://trendy-healthy-backend.store'; //서버 연결 후 사용할 APi
 
 const api = axios.create({
     baseURL: baseURL,
